fix(authenticators/oidc): prevent random logout
If a session is restored but the token is still valid, no refresh loop is started to prevent a
infinite token refresh loop on multiple tabs. This token is never checked again so we never start a
refresh loop even if the token is already exipred. This causes the next request to return a 401 and
invalidates the session.