Rework restore call
At the moment the restore call always makes a refresh request,
which then updates the session. A session update triggers the
restore call (if not done by this exact restore call) which leads
to a infinite request loop a soon as two or more tabs of the app
using this addon are open.
To fix this the restore call only makes a refresh request if the
token actually expired. Otherwise just schedule refresh which will
then make the actual refresh request.