feat: handle `attemptedTransition`
`ember-simple-auth` handles attempted transitions which are aborted by
the authentication process. Since the oidc addon redirects the page, the
attempted transition is lost and the user will always end up on the
index (or `routeAfterAuthentication`) page. This update will store the
attempted transition and recreate it after a successful login.