fix(token refresh): Don't lose token on refresh in network outage
When the token refresh fails for some reason, we retry for a given
amount of time. Since the retry happens in a timeout, we must take
care not to lose the token state when it finally succeeds again.

This moves the triggering of `sessionDataUpdated` to `_handleAuthResponse()`
so it cannot be forgotten accidentally, and takes care to always return
some data in _refresh (even if it's the old token)