fix(mixin): correctly recompute `headers` in the `oidc-adapter-mixin`
The computed property `headers` was watching the
`session.isAuthenticated` property which does not change when a
successful token refresh happend. This lead to an error since after the
refresh the `headers` property still returned the old token.
To fix this watch the `session.data.authenticated.id_token` as according
to the openid-connect specification
(https://openid.net/specs/openid-connect-core-1_0.html#TokenResponse)
the `id_token` must always be present.