feat(single-logout): separate session invalidate and oidc logout
There is a difference between a logout of the current application and a
logout on the authorization server. Until now the addon assumed if the
ember-simple-auth session is invalidated it should also call the
`end-session` endpoint of the authorization server which invalidates all
ongoing sessions in possible different applications. With the
introduction of the `singleLogout` function the application using this
addon can decide between invalidating the current ember-simple-auth
session or doing a global/single logout in addition.

BREAKING CHANGE: Since v1.0.0 this addon will always perform a single
logout on the authorization server. With this change the default
behaviour is "only" a logout on the current application. If the single
logout should be preserved the consuming application needs to manually
call the new `singleLogout` function.