fix(config): remove usage of ember-get-config
It's advised to always use `resolveRegistration` for getting the config
object if the owner is accessible from the code which it is in our case.
`ember-get-config` uses `@embroider/macros` to resolve the config which
can lead to unexpected content and duplicated config objects in the code
instead of the meta tag.