fix(apollo): make apollo client a dependency
This is necessary since embroider doesn't handle optional dependencies
properly as reported here: https://github.com/embroider-build/embroider/issues/1266

This commit makes `@apollo/client` temporarly a non optional dependency
to fix the current issue we have in apps without that dependency.
However, we need to revert this as soon as the issue above is resolved
since we don't want to enforce this dependency.

Related to #597