fix(config): fix configuration of `authEndpoint`
If using the standard configuration (without using auto discovery) the
`authEndpoint` was wrongly set to `undefined` in the config service
because it was overwritten by the value of `authorizationEndpoint` which
is not defined using the standard configuration pattern.

We now make sure to only apply aliases if there is a) a value to be
applied and b) no value present on the key to be written.

This fixes a regression introduced in #1122.